+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.55.0] - 2024-10-21
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Activation page at `/users/activate/[activationCode]`.
+- [Marek Kurańda](https://github.com/mjkuranda): Confirmation modal for all objects at manage page.
+- [Marek Kurańda](https://github.com/mjkuranda): Object count at every object manage page.
 
 ## [1.54.3] - 2024-10-21
 ### Fixed